refactor(ui): extract WordDisplay from TypingTestView.tsx (544→397 lines)
Extract WordDisplay component with charClassName, displayChar, Cursor
helpers and COMPOSITION_CHAR_CLASS constant into WordDisplay.tsx.
